+++ botocore/signers.py
@@ -90,7 +90,7 @@ class RequestSigner(object):
         return self.sign(operation_name, request)
 
     def sign(self, operation_name, request, region_name=None,
-             signing_type='standard', expires_in=None):
+             signing_type='standard', expires_in=None, signing_name=None):
         """Sign a request before it goes out over the wire.
 
         :type operation_name: string
@@ -113,9 +113,15 @@ class RequestSigner(object):
         :type expires_in: int
         :param expires_in: The number of seconds the presigned url is valid
             for. This parameter is only valid for signing type 'presign-url'.
+
+        :type signing_name: str
+        :param signing_name: The name to use for the service when signing.
         """
         if region_name is None:
             region_name = self._region_name
+
+        if signing_name is None:
+            signing_name = self._signing_name
 
         signature_version = self._choose_signer(
             operation_name, signing_type, request.context)
@@ -123,13 +129,13 @@ class RequestSigner(object):
         # Allow mutating request before signing
         self._event_emitter.emit(
             'before-sign.{0}.{1}'.format(self._service_name, operation_name),
-            request=request, signing_name=self._signing_name,
+            request=request, signing_name=signing_name,
             region_name=self._region_name,
             signature_version=signature_version, request_signer=self)
 
         if signature_version != botocore.UNSIGNED:
             kwargs = {
-                'signing_name': self._signing_name,
+                'signing_name': signing_name,
                 'region_name': region_name,
                 'signature_version': signature_version
             }
@@ -230,7 +236,8 @@ class RequestSigner(object):
     get_auth = get_auth_instance
 
     def generate_presigned_url(self, request_dict, operation_name,
-                               expires_in=3600, region_name=None):
+                               expires_in=3600, region_name=None,
+                               signing_name=None):
         """Generates a presigned url
 
         :type request_dict: dict
@@ -247,11 +254,14 @@ class RequestSigner(object):
         :type region_name: string
         :param region_name: The region name to sign the presigned url.
 
+        :type signing_name: str
+        :param signing_name: The name to use for the service when signing.
+
         :returns: The presigned url
         """
         request = create_request_object(request_dict)
         self.sign(operation_name, request, region_name,
-                  'presign-url', expires_in)
+                  'presign-url', expires_in, signing_name)
 
         request.prepare()
         return request.url
@@ -381,6 +391,61 @@ class CloudFrontSigner(object):
         # http://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/private-content-linux-openssl.html
         return base64.b64encode(
             data).replace(b'+', b'-').replace(b'=', b'_').replace(b'/', b'~')
+
+
+def add_generate_db_auth_token(class_attributes, **kwargs):
+    class_attributes['generate_db_auth_token'] = generate_db_auth_token
+
+
+def generate_db_auth_token(self, DBHostname, Port, DBUsername, Region=None):
+    """Generates an auth token used to connect to a db with IAM credentials.
+
+    :type DBHostname: str
+    :param DBHostname: The hostname of the database to connect to.
+
+    :type Port: int
+    :param Port: The port number the database is listening on.
+
+    :type DBUsername: str
+    :param DBUsername: The username to log in as.
+
+    :type Region: str
+    :param Region: The region the database is in. If None, the client
+        region will be used.
+
+    :return: A presigned url which can be used as an auth token.
+    """
+    region = Region
+    if region is None:
+        region = self.meta.region_name
+
+    params = {
+        'Action': 'connect',
+        'DBUser': DBUsername,
+    }
+
+    request_dict = {
+        'url_path': '/',
+        'query_string': '',
+        'headers': {},
+        'body': params,
+        'method': 'GET'
+    }
+
+    # RDS requires that the scheme not be set when sent over. This can cause
+    # issues when signing because the Python url parsing libraries follow
+    # RFC 1808 closely, which states that a netloc must be introduced by `//`.
+    # Otherwise the url is presumed to be relative, and thus the whole
+    # netloc would be treated as a path component. To work around this we
+    # introduce https here and remove it once we're done processing it.
+    scheme = 'https://'
+    endpoint_url = '%s%s:%s' % (scheme, DBHostname, Port)
+    prepare_request_dict(request_dict, endpoint_url)
+    presigned_url = self._request_signer.generate_presigned_url(
+        operation_name='connect', request_dict=request_dict,
+        region_name=region, expires_in=900, signing_name='rds-db'
+    )
+    return presigned_url[len(scheme):]
 
 
 class S3PostPresigner(object):