+++ tests/unit/test_signers.py
@@ -10,22 +10,24 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-
 import mock
 import datetime
 import json
 
+from dateutil.tz import tzutc
+
 import botocore
+import botocore.session
 import botocore.auth
 from botocore.compat import six, urlparse, parse_qs
-
+from botocore.config import Config
 from botocore.credentials import Credentials
 from botocore.credentials import ReadOnlyCredentials
 from botocore.exceptions import NoRegionError, UnknownSignatureVersionError
 from botocore.exceptions import UnknownClientMethodError, ParamValidationError
 from botocore.exceptions import UnsupportedSignatureVersionError
 from botocore.signers import RequestSigner, S3PostPresigner, CloudFrontSigner
-from botocore.config import Config
+from botocore.signers import generate_db_auth_token
 
 from tests import unittest
 
@@ -39,6 +41,29 @@ class BaseSignerTest(unittest.TestCase):
             'service_name', 'region_name', 'signing_name',
             'v4', self.credentials, self.emitter)
         self.fixed_credentials = self.credentials.get_frozen_credentials()
+
+    def _urlparse(self, url):
+        if isinstance(url, six.binary_type):
+            # Not really necessary, but it helps to reduce noise on Python 2.x
+            url = url.decode('utf8')
+        return urlparse(url)
+
+    def assert_url_equal(self, url1, url2):
+        parts1 = self._urlparse(url1)
+        parts2 = self._urlparse(url2)
+
+        # Because the query string ordering isn't relevant, we have to parse
+        # every single part manually and then handle the query string.
+        self.assertEqual(parts1.scheme, parts2.scheme)
+        self.assertEqual(parts1.netloc, parts2.netloc)
+        self.assertEqual(parts1.path, parts2.path)
+        self.assertEqual(parts1.params, parts2.params)
+        self.assertEqual(parts1.fragment, parts2.fragment)
+        self.assertEqual(parts1.username, parts2.username)
+        self.assertEqual(parts1.password, parts2.password)
+        self.assertEqual(parts1.hostname, parts2.hostname)
+        self.assertEqual(parts1.port, parts2.port)
+        self.assertEqual(parse_qs(parts1.query), parse_qs(parts2.query))
 
 
 class TestSigner(BaseSignerTest):
@@ -424,6 +449,43 @@ class TestSigner(BaseSignerTest):
             expires=2
         )
 
+    def test_sign_with_custom_signing_name(self):
+        request = mock.Mock()
+        auth = mock.Mock()
+        auth_types = {
+            'v4': auth
+        }
+        with mock.patch.dict(botocore.auth.AUTH_TYPE_MAPS, auth_types):
+            self.signer.sign('operation_name', request, signing_name='foo')
+        auth.assert_called_with(
+            credentials=ReadOnlyCredentials('key', 'secret', None),
+            service_name='foo',
+            region_name='region_name'
+        )
+
+    def test_presign_with_custom_signing_name(self):
+        auth = mock.Mock()
+        auth.REQUIRES_REGION = True
+
+        request_dict = {
+            'headers': {},
+            'url': 'https://foo.com',
+            'body': b'',
+            'url_path': '/',
+            'method': 'GET',
+            'context': {}
+        }
+        with mock.patch.dict(botocore.auth.AUTH_TYPE_MAPS,
+                             {'v4-query': auth}):
+            presigned_url = self.signer.generate_presigned_url(
+                request_dict, operation_name='operation_name',
+                signing_name='foo')
+        auth.assert_called_with(
+            credentials=self.fixed_credentials,
+            region_name='region_name',
+            expires=3600, service_name='foo')
+        self.assertEqual(presigned_url, 'https://foo.com')
+
     def test_unknown_signer_raises_unknown_on_standard(self):
         request = mock.Mock()
         auth = mock.Mock()
@@ -453,8 +515,9 @@ class TestSigner(BaseSignerTest):
                                  signing_type='presign-post')
 
 
-class TestCloudfrontSigner(unittest.TestCase):
+class TestCloudfrontSigner(BaseSignerTest):
     def setUp(self):
+        super(TestCloudfrontSigner, self).setUp()
         self.signer = CloudFrontSigner("MY_KEY_ID", lambda message: b'signed')
         # It helps but the long string diff will still be slightly different on
         # Python 2.6/2.7/3.x. We won't soly rely on that anyway, so it's fine.
@@ -485,20 +548,6 @@ class TestCloudfrontSigner(unittest.TestCase):
         }
         self.assertEqual(json.loads(policy), expected)
 
-    def _urlparse(self, url):
-        if isinstance(url, six.binary_type):
-            # Not really necessary, but it helps to reduce noise on Python 2.x
-            url = url.decode('utf8')
-        return dict(urlparse(url)._asdict())  # Needs an unordered dict here
-
-    def assertEqualUrl(self, url1, url2):
-        # We compare long urls by their dictionary parts
-        parts1 = self._urlparse(url1)
-        parts2 = self._urlparse(url2)
-        self.assertEqual(
-            parse_qs(parts1.pop('query')), parse_qs(parts2.pop('query')))
-        self.assertEqual(parts1, parts2)
-
     def test_generate_presign_url_with_expire_time(self):
         signed_url = self.signer.generate_presigned_url(
             'http://test.com/foo.txt',
@@ -506,7 +555,7 @@ class TestCloudfrontSigner(unittest.TestCase):
         expected = (
             'http://test.com/foo.txt?Expires=1451606400&Signature=c2lnbmVk'
             '&Key-Pair-Id=MY_KEY_ID')
-        self.assertEqualUrl(signed_url, expected)
+        self.assert_url_equal(signed_url, expected)
 
     def test_generate_presign_url_with_custom_policy(self):
         policy = self.signer.build_policy(
@@ -523,7 +572,7 @@ class TestCloudfrontSigner(unittest.TestCase):
             '6IjEyLjM0LjU2Ljc4LzkifSwiRGF0ZUdyZWF0ZXJUaGFuIjp7IkFX'
             'UzpFcG9jaFRpbWUiOjE0NDg5MjgwMDB9fX1dfQ__'
             '&Signature=c2lnbmVk&Key-Pair-Id=MY_KEY_ID')
-        self.assertEqualUrl(signed_url, expected)
+        self.assert_url_equal(signed_url, expected)
 
 
 class TestS3PostPresigner(BaseSignerTest):
@@ -857,3 +906,46 @@ class TestGeneratePresignedPost(unittest.TestCase):
         self.client = self.session.create_client('ec2', 'us-west-2')
         with self.assertRaises(AttributeError):
             self.client.generate_presigned_post()
+
+
+class TestGenerateDBAuthToken(BaseSignerTest):
+    def setUp(self):
+        self.session = botocore.session.get_session()
+        self.client = self.session.create_client(
+            'rds', region_name='us-east-1', aws_access_key_id='akid',
+            aws_secret_access_key='skid', config=Config(signature_version='v4')
+        )
+
+    def test_generate_db_auth_token(self):
+        hostname = 'prod-instance.us-east-1.rds.amazonaws.com'
+        port = 3306
+        username = 'someusername'
+        clock = datetime.datetime(2016, 11, 7, 17, 39, 33, tzinfo=tzutc())
+
+        with mock.patch('datetime.datetime') as dt:
+            dt.utcnow.return_value = clock
+            result = generate_db_auth_token(
+                self.client, hostname, port, username)
+
+        expected_result = (
+            'prod-instance.us-east-1.rds.amazonaws.com:3306/?Action=connect'
+            '&DBUser=someusername&X-Amz-Algorithm=AWS4-HMAC-SHA256'
+            '&X-Amz-Date=20161107T173933Z&X-Amz-SignedHeaders=host'
+            '&X-Amz-Expires=900&X-Amz-Credential=akid%2F20161107%2F'
+            'us-east-1%2Frds-db%2Faws4_request&X-Amz-Signature'
+            '=d1138cdbc0ca63eec012ec0fc6c2267e03642168f5884a7795320d4c18374c61'
+        )
+        self.assert_url_equal(result, expected_result)
+
+    def test_custom_region(self):
+        hostname = 'host.us-east-1.rds.amazonaws.com'
+        port = 3306
+        username = 'mySQLUser'
+        region = 'us-west-2'
+        result = generate_db_auth_token(
+            self.client, hostname, port, username, Region=region)
+
+        self.assertIn(region, result)
+        # The hostname won't be changed even if a different region is specified
+        self.assertIn(hostname, result)
+